create a initial structure of models class
