+++ service/TrainAndEvaluateService.py
@@ -20,7 +20,7 @@ from models.dialogue_policy.supervised_learning.TedPolicy import Ted
 from models.dialogue_policy.supervised_learning.LedPolicy import Led
 from pytorch_lightning.callbacks.early_stopping import EarlyStopping
 from service.InputOutput.MongoDB import MongoDB
-from models.dialogue_policy.supervised_learning.StarSpacePolicy import StarSpacePolicy
+from models.dialogue_policy.supervised_learning.EmbeddingPolicy import EmbeddingPolicy
 from view.Logger import Logger
 
 
@@ -102,7 +102,7 @@ class TrainAndEvaluateService(Pipeline):
         models = {
             "TED": Ted,
             "LED": Led,
-            "SS": StarSpacePolicy
+            "SS": EmbeddingPolicy
         }
         return models[model](config, actions)
